src/ma_finkg/requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
pydantic>=2.0.0
typing-extensions>=4.0.0
json-repair>=0.7.0
pytest>=7.0.0
pyyaml>=6.0.0